Reorganize tqdm modules architecture to avoid unnecessary imports (no imports interaction fix #176 #245 #188 + speed boost fix #258)
