+++ tqdm/_main.py
@@ -1,4 +1,4 @@
-from ._tqdm import tqdm, TqdmTypeError, TqdmKeyError
+from .core import tqdm, TqdmTypeError, TqdmKeyError
 from ._version import __version__  # NOQA
 import sys
 import re
